Refactor logger implementation and remove Python dependency
- Deleted obsolete Python logger script (logger.py).
- Introduced a new Zig-based logging system with a LoggerAllocator for thread-unsafe logging.
- Implemented JSON parsing and formatting for log entries.
- Created a build script (build.zig) for managing the Zig project.
- Added unit tests for logging functionality and JSON output generation.
- Removed unnecessary cache files and updated project structure.
- Added configuration for maximum bytes in log files.
- Introduced structured logging with log levels and entries.
