--- .zig-cache/o/ebd7ddab8ffe003267120d598aecce68/dependencies.zig
@@ -1,2 +1,0 @@
-pub const packages = struct {};
-pub const root_deps: []const struct { []const u8, []const u8 } = &.{};
