+++ core/src/config.zig
@@ -1,0 +1,1 @@
+pub const MAX_BYTES = 10 * 1024 * 1024; // 10 MB
