--- logger_output.json
@@ -1,24 +1,0 @@
-{
-  "API_ROOT": {
-    "001": {
-      "level": "Debug",
-      "message": "debug message"
-    },
-    "002": {
-      "level": "Info",
-      "message": "info message"
-    },
-    "003": {
-      "level": "Warning",
-      "message": "warning message"
-    },
-    "004": {
-      "level": "Error",
-      "message": "error message"
-    },
-    "005": {
-      "level": "Critical",
-      "message": "critical message"
-    }
-  }
-}
